src/main/java/yapp/buddycon/app/gifticon/application/service/GifticonService.java
package yapp.buddycon.app.gifticon.application.service;

import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;
import yapp.buddycon.app.gifticon.adapter.request.SearchGifticonDTO;
import yapp.buddycon.app.gifticon.adapter.response.GifticonResponseDTO;
import yapp.buddycon.app.gifticon.application.port.in.GifticonUseCase;
import yapp.buddycon.app.gifticon.application.port.out.GifticonQueryPort;

@Service
@RequiredArgsConstructor
@Transactional(readOnly = true)
public class GifticonService implements GifticonUseCase {

  private final GifticonQueryPort gifticonQueryPort;

  @Override
  public Page<GifticonResponseDTO> getGifticons(SearchGifticonDTO dto) {
    return gifticonQueryPort.findAll(dto, PageRequest.of(dto.getPageNumber(), dto.getRowCount()));
  }

}
